Subo la entrega final de JS usando API y librerias, borrando paginas de DW que no usaba JS

+++ JS/Carrito.js
@@ -1,132 +1,30 @@
-
-    /* document.addEventListener("DOMContentLoaded", () => {
-        
-        const carritoProductos = JSON.parse(localStorage.getItem("productoEnCarrito")) || [];
-        
-        
-        let carritoVacio = document.querySelector("#carritoVacio");
-        let contenedorCarritoProductos = document.querySelector("#carritoProductos");
-        let carritoAcciones = document.querySelector("#carritoAcciones");
-        let botonBorrar = document.querySelectorAll(".delete")
-    
-        
-        if (carritoProductos.length > 0) {
-           
-            carritoVacio.classList.add("disabled");
-            contenedorCarritoProductos.classList.remove("disabled");
-            carritoAcciones.classList.remove("disabled");
-    
-            const table = document.createElement("table");
-            table.className = "cartTable offset-lg-2 col-lg-8";
-    
-            const tbody = document.createElement("tbody");
-    
-            // Fila de encabezado
-            const headerRow = document.createElement("tr");
-            headerRow.className = "CartProducto";
-    
-            const headers = ["Producto", "Producto", "Precio", "Cant.", "SubTotal", ""];
-            headers.forEach(headerText => {
-                const th = document.createElement("td");
-                th.textContent = headerText;
-                headerRow.appendChild(th);
-                
-            });
-    
-            tbody.appendChild(headerRow);
-    
-            // Iterar sobre productos del carrito y agregarlos a la tabla
-            carritoProductos.forEach(producto => {
-                const filaProductos = document.createElement("tr");
-                filaProductos.className = "CartProducto";
-    
-                const imgTd = document.createElement("td");
-                imgTd.className = "imgProductoCarrito";
-                const img = document.createElement("img");
-                img.src = producto.img; // Asegúrate de que `producto.imagen` contenga la URL de la imagen
-                imgTd.appendChild(img);
-    
-                const descripcionTd = document.createElement("td");
-                descripcionTd.className = "descripcionProductoCarrito";
-                descripcionTd.innerHTML = `
-                    <div class="CartDescription">
-                        <p class="descripcionProductoEspecifico">${producto.nombre}</p>
-                        <p class="codigoNum">Cod. ${producto.id}</p>
-                    </div>
-                `;
-    
-                const precioTd = document.createElement("td");
-                precioTd.className = "productoPrecio";
-                precioTd.innerHTML = `<p class="productoPrecioCarrito">$ ${producto.precio}</p>`;
-    
-                const cantidadTd = document.createElement("td");
-                cantidadTd.className = "car-cant";
-                cantidadTd.innerHTML = `
-                    <input class="cantidadInput" type="number" style="width:50px; text-align:center" min="0" value="${producto.cantidad}">
-                `;
-    
-                const subtotalTd = document.createElement("td");
-                subtotalTd.className = "subtotalNumero";
-                subtotalTd.innerHTML = `$ ${producto.precio * producto.cantidad}`;
-    
-                const deleteTd = document.createElement("td");
-                deleteTd.className = "delete";
-                deleteTd.innerHTML = '<button><i class="bi bi-trash"></i></button>';
-    
-                filaProductos.appendChild(imgTd);
-                filaProductos.appendChild(descripcionTd);
-                filaProductos.appendChild(precioTd);
-                filaProductos.appendChild(cantidadTd);
-                filaProductos.appendChild(subtotalTd);
-                filaProductos.appendChild(deleteTd);
-    
-                tbody.appendChild(filaProductos);
-            });
-            actualizarBotonesEliminar()
-    
-            
-            table.appendChild(tbody);
-    
-           
-            contenedorCarritoProductos.appendChild(table);
-        }
-        
-    });
-    
-    function actualizarNumero(arrayDeProdCarrito) {
-        let nuevoNumero = arrayDeProdCarrito.reduce((acc, producto) => acc + producto.cantidad, 0);
-        numeroCarritoID.innerText = nuevoNumero; 
-    }
-   
-    function actualizarBotonesEliminar() {
-        botonBorrar = document.querySelectorAll(".delete");
-        botonBorrar.forEach(boton => {
-            boton.addEventListener("click", eliminarDelCarrito)
-        })
-    }
-   
-function eliminarDelCarrito(e){
-        let idBoton = e.currentTarget.id
-        let productoEliminado = carritoProductos.find(producto=>producto.id===idBoton)
-    
-        
-} */
 let carritoVacio = document.querySelector("#carritoVacio");
 let contenedorCarritoProductos = document.querySelector("#carritoProductos");
 let carritoAcciones = document.querySelector("#carritoAcciones");
 let botonBorrar = document.querySelectorAll(".delete button")
 let botonComprar = document.querySelector(".comprar")
-let contenedorParrafoCompra=document.querySelector(".contenedorCarrito")
+let contenedorParrafoCompra = document.querySelector(".contenedorCarrito")
+
+async function obtenerCotizacionDolar() {
+    try {
+        const response = await fetch("https://dolarapi.com/v1/dolares/oficial");
+        const data = await response.json();
+        return data.venta;
+    } catch (error) {
+        console.error('Error al obtener la cotización del dólar:', error);
+        return null;
+    }
+}
 
 
-document.addEventListener("DOMContentLoaded", function() {
+document.addEventListener("DOMContentLoaded", function () {
     const carritoProductos = JSON.parse(localStorage.getItem("productoEnCarrito")) || [];
 
     if (carritoProductos.length > 0) {
         carritoVacio.classList.add("disabled");
         contenedorCarritoProductos.classList.remove("disabled");
         carritoAcciones.classList.remove("disabled");
-        
+
 
         const table = document.createElement("table");
         table.className = "cartTable offset-lg-2 col-lg-8";
@@ -146,7 +44,7 @@ document.addEventListener("DOMContentLoaded", function() {
 
         tbody.appendChild(headerRow);
 
-        // Iterar sobre productos del carrito y agregarlos a la tabla
+
         carritoProductos.forEach(producto => {
             const filaProductos = document.createElement("tr");
             filaProductos.className = "CartProducto";
@@ -154,7 +52,7 @@ document.addEventListener("DOMContentLoaded", function() {
             const imgTd = document.createElement("td");
             imgTd.className = "imgProductoCarrito";
             const img = document.createElement("img");
-            img.src = producto.img; // Asegúrate de que `producto.img` contenga la URL de la imagen
+            img.src = producto.img;
             imgTd.appendChild(img);
 
             const descripcionTd = document.createElement("td");
@@ -173,7 +71,7 @@ document.addEventListener("DOMContentLoaded", function() {
             const cantidadTd = document.createElement("td");
             cantidadTd.className = "car-cant";
             cantidadTd.innerHTML = `
-                <input class="cantidadInput" type="number" style="width:50px; text-align:center" min="0" value="${producto.cantidad}">
+            <p class="productoPrecioCarrito">${producto.cantidad}</p>
             `;
 
             const subtotalTd = document.createElement("td");
@@ -184,12 +82,8 @@ document.addEventListener("DOMContentLoaded", function() {
             deleteTd.className = "delete";
             deleteTd.innerHTML = '<button id="' + producto.id + '"><i class="bi bi-trash"></i></button>';
 
-            filaProductos.appendChild(imgTd);
-            filaProductos.appendChild(descripcionTd);
-            filaProductos.appendChild(precioTd);
-            filaProductos.appendChild(cantidadTd);
-            filaProductos.appendChild(subtotalTd);
-            filaProductos.appendChild(deleteTd);
+            filaProductos.append(imgTd, descripcionTd, precioTd, cantidadTd, subtotalTd, deleteTd);
+
 
             tbody.appendChild(filaProductos);
         });
@@ -199,10 +93,11 @@ document.addEventListener("DOMContentLoaded", function() {
 
         actualizarBotonesEliminar();
         Actualizartotal()
-        
+
     }
 
-;})
+    ;
+})
 
 
 
@@ -211,59 +106,82 @@ function actualizarBotonesEliminar() {
     botonBorrar = document.querySelectorAll(".delete button");
     botonBorrar.forEach(boton => {
         boton.addEventListener("click", eliminarDelCarrito);
-        
+
     });
 }
 
 function eliminarDelCarrito(e) {
     let idBoton = e.currentTarget.id;
-    
-    let carritoProductos = JSON.parse(localStorage.getItem("productoEnCarrito")) || [];
-    let productoEliminado = carritoProductos.find(producto => producto.id === idBoton);
-    
-    if (productoEliminado) {
-        
-        carritoProductos = carritoProductos.filter(producto => producto.id !== idBoton);
-        localStorage.setItem("productoEnCarrito", JSON.stringify(carritoProductos));
-        
+    const filaProducto = e.currentTarget.closest('tr');
 
-        
-        const filaProducto = e.currentTarget.closest('tr');
-        filaProducto.remove();
+    Swal.fire({
+        title: "Estas seguro de eliminar esta linea de productos?",
+        text: "Podras volver agregarlo de nuevo desde Productos",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, quiero borrarlo!"
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire({
+                title: "Borrado",
+                text: "Este producto se elimino del carrito",
+                icon: "success"
+            });
 
-        
-        Actualizartotal()
-        actualizarNumero()
-    }
-    
+            let carritoProductos = JSON.parse(localStorage.getItem("productoEnCarrito")) || [];
+            let productoEliminado = carritoProductos.find(producto => producto.id === idBoton);
+
+            if (productoEliminado) {
+                carritoProductos = carritoProductos.filter(producto => producto.id !== idBoton);
+                localStorage.setItem("productoEnCarrito", JSON.stringify(carritoProductos));
+
+                filaProducto && filaProducto.remove();
+
+                Actualizartotal();
+                actualizarNumero();
+            }
+            if(carritoProductos.length==0){
+                carritoVacio.classList.remove("disabled");
+                contenedorCarritoProductos.classList.add("disabled");
+                carritoAcciones.classList.add("disabled");
+            }
+        }
+    });
 }
+
+
 function actualizarNumero() {
     let carritoProductos = JSON.parse(localStorage.getItem("productoEnCarrito")) || [];
     let nuevoNumero = carritoProductos.reduce((acc, producto) => acc + producto.cantidad, 0);
-    numeroCarritoID.innerText = nuevoNumero; 
+    numeroCarritoID.innerText = nuevoNumero;
 }
 
 
 actualizarNumero()
 
-function Actualizartotal(){
+async function Actualizartotal() {
     let valorTotalp = document.querySelector(".valorTotal")
+    const cotizacionDolar = await obtenerCotizacionDolar();
+
     let carritoProductos = JSON.parse(localStorage.getItem("productoEnCarrito")) || [];
-    const totalCalculado = carritoProductos.reduce((acc , producto) => acc + (producto.precio * producto.cantidad), 0)
-    valorTotalp.innerText= `$${totalCalculado}`
+    const precioEnDolares = carritoProductos.reduce((acc, producto) => acc + ((producto.precio * producto.cantidad) / cotizacionDolar), 0)
+    const precioEnDolaresDecimal = precioEnDolares.toFixed(2)
+    const totalCalculado = carritoProductos.reduce((acc, producto) => acc + (producto.precio * producto.cantidad), 0)
+    valorTotalp.innerText = `$${totalCalculado} / USD ${precioEnDolaresDecimal}`
 }
 
 botonComprar.addEventListener("click", comprarCarrito)
-function comprarCarrito(){
-    carritoProductos.length=0
-    localStorage.setItem("productosEnCarrito", JSON.stringify(carritoProductos))
-    
+
+function comprarCarrito() {
+    carritoProductos.length = 0
+    localStorage.clear()
     carritoVacio.classList.add("disabled");
     contenedorCarritoProductos.classList.add("disabled");
     carritoAcciones.classList.add("disabled");
-    let parrafoCompra= document.createElement("p")
-    parrafoCompra.innerText= "Gracias por su compra"
+    let parrafoCompra = document.createElement("p")
+    parrafoCompra.innerText = "Gracias por su compra"
     contenedorParrafoCompra.appendChild(parrafoCompra)
-
-    }
-
+    actualizarNumero()
+}
